Split EditWidgetsOverlay.kt into 8 files by responsibility
Pure move, no behavior change - follows the same precedent as
commit d6995d7's SettingsContent.kt split. 1845 lines down to 667
in the retained file (main composable, PlaceholderWidgetBox,
widgetCatalogFor, StateGroup.displayLabel); the rest moves to
7 new same-package files:

- WidgetDragGeometry.kt: drag-to-move math (dragMaxCell, DragStep,
  nextDragCell)
- WidgetResizeGeometry.kt: drag-to-resize math and the ResizeHandle
  composable
- WidgetTypeDialogs.kt: AddWidgetDialog, ConfigureWidgetDialog,
  WidgetType/MediaType label helpers
- WidgetImageConfigControls.kt: CustomImageConfig, ScaleModeConfig,
  ImageEffectsConfig, PanZoomConfig, FallbackArtworkConfig
- WidgetTransitionConfigControls.kt: ImageTransitionPicker,
  LogoTransitionPicker, GlintConfig
- WidgetColorConfig.kt: ColorBackgroundConfig, ColorSwatchPicker,
  HexColorInput, hex color helpers
- WidgetTextConfigControls.kt: GameDescriptionConfig, RatingConfig

Several composables/values promoted from private to internal where
a different new file now calls them across the split (e.g.
ConfigureWidgetDialog calling into the per-type config composables,
GameDescriptionConfig sharing COLOR_PRESETS/HexColorInput with
WidgetColorConfig.kt) - all same-package, so no new imports needed.
EditWidgetsOverlayTest.kt required zero changes, since it already
relied on the same-package internal visibility these functions had
before the split.

detekt baseline entries for the moved code were derived directly
from detekt's own output against the split files (not hand-guessed)
to avoid mismatched signature text, and the stale whole-file
TooManyFunctions entry was dropped outright since no individual new
file trips that threshold.
