+++ app/src/main/java/com/esde/companion/ui/widgets/edit/EditWidgetsOverlay.kt
@@ -12,60 +12,30 @@ import androidx.compose.animation.fadeOut
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.gestures.detectDragGestures
 import androidx.compose.foundation.gestures.detectTapGestures
-import androidx.compose.foundation.horizontalScroll
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxWithConstraints
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.ArrowForward
-import androidx.compose.material.icons.filled.Block
-import androidx.compose.material.icons.filled.BlurOn
 import androidx.compose.material.icons.filled.Check
-import androidx.compose.material.icons.filled.Crop
-import androidx.compose.material.icons.filled.CropFree
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.FlipToBack
 import androidx.compose.material.icons.filled.FlipToFront
 import androidx.compose.material.icons.filled.MoreVert
 import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material.icons.filled.ZoomIn
-import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.SegmentedButton
-import androidx.compose.material3.SegmentedButtonDefaults
-import androidx.compose.material3.SingleChoiceSegmentedButtonRow
-import androidx.compose.material3.Slider
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
@@ -78,8 +48,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.hapticfeedback.HapticFeedbackType
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.platform.LocalDensity
@@ -92,32 +60,17 @@ import androidx.compose.ui.zIndex
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.esde.companion.data.storage.SafPathResolver
 import com.esde.companion.domain.model.GridDimensions
-import com.esde.companion.domain.model.ImageEffects
-import com.esde.companion.domain.model.ImageTransitionMode
-import com.esde.companion.domain.model.LogoTransitionMode
 import com.esde.companion.domain.model.MediaType
-import com.esde.companion.domain.model.NoRatingBehavior
 import com.esde.companion.domain.model.PlacedWidget
 import com.esde.companion.domain.model.ScaleMode
 import com.esde.companion.domain.model.StateGroup
 import com.esde.companion.domain.model.WidgetContent
 import com.esde.companion.domain.model.WidgetType
-import com.esde.companion.domain.model.allowsFadeTransition
-import com.esde.companion.domain.model.fallbackMediaTypeOptions
-import com.esde.companion.domain.model.glintEnabled
-import com.esde.companion.domain.model.imageTransitionMode
-import com.esde.companion.domain.model.isLogoStyle
-import com.esde.companion.domain.model.logoTransitionMode
-import com.esde.companion.domain.model.supportsFallbackArtwork
-import com.esde.companion.domain.model.supportsImageTransition
-import com.esde.companion.domain.model.supportsPanZoom
 import com.esde.companion.ui.CORNER_BUTTON_EDGE_PADDING
 import com.esde.companion.ui.CornerFab
 import com.esde.companion.ui.dock.DockPreview
 import com.esde.companion.ui.widgets.WidgetContentView
 import com.esde.companion.ui.widgets.gridDimensionsFor
-import java.util.Locale
-import kotlin.math.roundToInt
 
 /** Options button opacity at rest - translucent so it's unobtrusive but still visible. */
 private const val OPTIONS_BUTTON_IDLE_ALPHA = 0.75f
@@ -126,28 +79,11 @@ private const val OPTIONS_BUTTON_IDLE_ALPHA = 0.75f
  * distracts from placement/sizing happening elsewhere on screen. */
 private const val OPTIONS_BUTTON_DRAGGING_ALPHA = 0.15f
 
-/** A widget can never be resized smaller than one grid cell in either dimension. */
-private const val MIN_SPAN = 1
-
-/** Touch target size for a resize handle - offsets/positioning throughout this file are
- * all in terms of this, so the grab zone stays full-size regardless of [HANDLE_DOT_SIZE]. */
-private val HANDLE_SIZE = 32.dp
-
-/** Visual size of the dot drawn inside a resize handle's touch target - smaller than
- * [HANDLE_SIZE] so the dot reads as a compact grab indicator without shrinking the
- * actual (harder-to-hit-precisely) touch target to match. */
-private val HANDLE_DOT_SIZE = 24.dp
-
 /** How far PlaceholderWidgetBox insets its visible border from the widget's true grid
  * bounds (its own `.padding(2.dp)`) - resize handle offsets subtract this so a handle
  * centers on the border as it's actually drawn, not on the untouched cell boundary 2dp
  * further out. Kept as one named constant so the two can't silently drift apart. */
 private val WIDGET_BORDER_INSET = 2.dp
-
-/** Once the resize handle's actual on-screen position gets this close to the grid's true
- * edge, snap straight to the grid boundary rather than requiring further delta-based
- * drag distance - see ResizeHandle's kdoc for why this is needed at all. */
-private val EDGE_SNAP_THRESHOLD = 24.dp
 
 private val MENU_SHAPE = RoundedCornerShape(16.dp)
 
@@ -291,6 +227,7 @@ fun EditWidgetsOverlay(
                         },
             ) {
                 val grid = remember(maxWidth, maxHeight) { gridDimensionsFor(maxWidth, maxHeight) }
+
                 // Switch to whichever canvas edit mode was entered for, before the first
                 // load - selectCanvas() no-ops if it's already the current canvas (the
                 // common case, since it defaults to System), and this runs before
@@ -583,57 +520,6 @@ fun EditWidgetsOverlay(
     }
 }
 
-/** The furthest column/row a widget with [span] cells can be dragged to within a grid
- * spanning [gridExtent] cells. coerceAtLeast(0) keeps callers' coerceIn from throwing on
- * an empty range when a saved widget's span exceeds a live-measured smaller grid -
- * pinning to the origin instead of crashing. */
-internal fun dragMaxCell(
-    gridExtent: Int,
-    span: Int,
-): Int = (gridExtent - span).coerceAtLeast(0)
-
-/** The result of feeding one drag tick's delta into [nextDragCell]. */
-internal data class DragStep(
-    val cell: IntOffset,
-    val accum: Offset,
-    val moved: Boolean,
-    val hitBoundary: Boolean,
-)
-
-/** Grid-snapped drag-to-move math for a single [detectDragGestures] tick: accumulates
- * sub-cell drag distance in [accum] until it crosses a full cell (sized [cellSizePx]),
- * then advances [currentCell] by that many cells (clamped to [maxCell]), carrying the
- * leftover sub-cell remainder forward rather than discarding it. [DragStep.hitBoundary]
- * is true only on the tick a coordinate transitions onto a boundary, not on every
- * subsequent tick spent pinned against it. */
-internal fun nextDragCell(
-    dragAmount: Offset,
-    accum: Offset,
-    cellSizePx: Offset,
-    currentCell: IntOffset,
-    maxCell: IntOffset,
-): DragStep {
-    val newAccum = accum + dragAmount
-    val columnDelta = (newAccum.x / cellSizePx.x).toInt()
-    val rowDelta = (newAccum.y / cellSizePx.y).toInt()
-
-    if (columnDelta == 0 && rowDelta == 0) {
-        return DragStep(currentCell, newAccum, moved = false, hitBoundary = false)
-    }
-
-    val newColumn = (currentCell.x + columnDelta).coerceIn(0, maxCell.x)
-    val newRow = (currentCell.y + rowDelta).coerceIn(0, maxCell.y)
-    val hitBoundary =
-        (newColumn != currentCell.x && (newColumn == 0 || newColumn == maxCell.x)) ||
-            (newRow != currentCell.y && (newRow == 0 || newRow == maxCell.y))
-    return DragStep(
-        cell = IntOffset(newColumn, newRow),
-        accum = Offset(newAccum.x - columnDelta * cellSizePx.x, newAccum.y - rowDelta * cellSizePx.y),
-        moved = true,
-        hitBoundary = hitBoundary,
-    )
-}
-
 /**
  * Grid-snapped drag-to-move, plus tap-to-select (shows the resize handle, see
  * ResizeHandle). Tracks its own cumulative logical position and sub-cell drag remainder
@@ -779,1067 +665,3 @@ private fun PlaceholderWidgetBox(
         }
     }
 }
-
-/** The largest span a resize handle on [isFarEdge] can grow a widget to, given the
- * anchored [start] cell, current [span], and a grid spanning [gridExtentCells] cells.
- * coerceAtLeast(MIN_SPAN) keeps callers' coerceIn from throwing on an empty range when a
- * saved widget's position leaves less room than MIN_SPAN on a live-measured smaller grid. */
-internal fun resizeMaxSpan(
-    gridExtentCells: Int,
-    start: Int,
-    span: Int,
-    isFarEdge: Boolean,
-): Int =
-    if (isFarEdge) {
-        (gridExtentCells - start).coerceAtLeast(MIN_SPAN)
-    } else {
-        (start + span).coerceAtLeast(MIN_SPAN)
-    }
-
-/** A resize handle's position/size state: the anchored [start] cell, current [span], and
- * [maxSpan] it can grow to (see [resizeMaxSpan]). */
-internal data class ResizeBounds(
-    val start: Int,
-    val span: Int,
-    val maxSpan: Int,
-)
-
-/** The on-screen pixel geometry a resize handle needs to detect edge proximity: the size
- * of one grid cell, the grid's total extent, how close counts as "near" ([edgeSnapThresholdPx]),
- * and the gesture's starting pixel position for the near/far edge respectively. */
-internal data class ResizeGeometry(
-    val cellPx: Float,
-    val gridExtentPx: Float,
-    val edgeSnapThresholdPx: Float,
-    val initialStartPx: Float,
-    val initialFarEdgePx: Float,
-)
-
-/** A resize gesture's running sub-cell drag remainder ([accum], reset on each whole-cell
- * step) and total on-screen displacement since the gesture began ([totalDrag], never
- * reset - see [nextResizeSpan]). */
-internal data class ResizeAccumulator(
-    val accum: Float,
-    val totalDrag: Float,
-)
-
-/** The result of feeding one drag tick's delta into [nextResizeSpan]. [changed] is true
- * whenever [start]/[span] moved, whether by a whole-cell step or an edge snap;
- * [snapped] narrows that to just the edge-snap case, which fires its own haptic. */
-internal data class ResizeStep(
-    val start: Int,
-    val span: Int,
-    val accumulator: ResizeAccumulator,
-    val changed: Boolean,
-    val snapped: Boolean,
-)
-
-/** Grid-snapped drag-to-resize math for a single [detectDragGestures] tick, covering both
- * a far-edge handle (Right/Bottom: [ResizeBounds.start] fixed, span grows/shrinks away
- * from it) and a near-edge handle (Left/Top: span's far edge fixed, start moves and span
- * compensates). Sub-cell [delta] accumulates until it crosses a full cell, same mechanism
- * as [nextDragCell]. [ResizeAccumulator.totalDrag] is the gesture's total on-screen
- * displacement - once it puts the moving edge within [ResizeGeometry.edgeSnapThresholdPx]
- * of the grid boundary, this snaps straight to the limit rather than requiring the exact
- * remaining per-cell delta to accumulate, since delta-based dragging alone can never reach
- * a physical screen edge (no more room for the finger to keep moving). */
-internal fun nextResizeSpan(
-    isFarEdge: Boolean,
-    delta: Float,
-    accumulator: ResizeAccumulator,
-    bounds: ResizeBounds,
-    geometry: ResizeGeometry,
-): ResizeStep {
-    val newAccumulator = ResizeAccumulator(accumulator.accum + delta, accumulator.totalDrag + delta)
-    val resolved =
-        if (isFarEdge) {
-            resolveFarEdgeResize(bounds, geometry, newAccumulator)
-        } else {
-            resolveNearEdgeResize(bounds, geometry, newAccumulator)
-        }
-    val resolvedAccumulator = newAccumulator.copy(accum = resolved.accum)
-    return ResizeStep(resolved.start, resolved.span, resolvedAccumulator, resolved.changed, resolved.snapped)
-}
-
-private data class ResolvedResize(
-    val start: Int,
-    val span: Int,
-    val accum: Float,
-    val changed: Boolean,
-    val snapped: Boolean,
-)
-
-/** Right/Bottom: the far edge moves with the drag, [ResizeBounds.start] is fixed. */
-private fun resolveFarEdgeResize(
-    bounds: ResizeBounds,
-    geometry: ResizeGeometry,
-    accumulator: ResizeAccumulator,
-): ResolvedResize {
-    val farEdgePx = geometry.initialFarEdgePx + accumulator.totalDrag
-    val nearGridBoundary = geometry.gridExtentPx - farEdgePx <= geometry.edgeSnapThresholdPx
-    val cellDelta = if (nearGridBoundary) 0 else (accumulator.accum / geometry.cellPx).toInt()
-    return when {
-        nearGridBoundary && bounds.span != bounds.maxSpan ->
-            ResolvedResize(bounds.start, bounds.maxSpan, accumulator.accum, changed = true, snapped = true)
-        cellDelta != 0 ->
-            ResolvedResize(
-                bounds.start,
-                (bounds.span + cellDelta).coerceIn(MIN_SPAN, bounds.maxSpan),
-                accumulator.accum - cellDelta * geometry.cellPx,
-                changed = true,
-                snapped = false,
-            )
-        else -> ResolvedResize(bounds.start, bounds.span, accumulator.accum, changed = false, snapped = false)
-    }
-}
-
-/** Left/Top: the near edge (start) moves with the drag, the far edge is fixed - start and
- * span must update together. */
-private fun resolveNearEdgeResize(
-    bounds: ResizeBounds,
-    geometry: ResizeGeometry,
-    accumulator: ResizeAccumulator,
-): ResolvedResize {
-    val startPx = geometry.initialStartPx + accumulator.totalDrag
-    val nearGridBoundary = startPx <= geometry.edgeSnapThresholdPx
-    val cellDelta = if (nearGridBoundary) 0 else (accumulator.accum / geometry.cellPx).toInt()
-    return when {
-        nearGridBoundary && bounds.start != 0 ->
-            ResolvedResize(0, bounds.maxSpan, accumulator.accum, changed = true, snapped = true)
-        cellDelta != 0 -> {
-            val newStart = (bounds.start + cellDelta).coerceIn(0, bounds.maxSpan - MIN_SPAN)
-            ResolvedResize(
-                newStart,
-                bounds.maxSpan - newStart,
-                accumulator.accum - cellDelta * geometry.cellPx,
-                changed = true,
-                snapped = false,
-            )
-        }
-        else -> ResolvedResize(bounds.start, bounds.span, accumulator.accum, changed = false, snapped = false)
-    }
-}
-
-/** Which side of the selected widget a [ResizeHandle] sits on and drags along. Left/Right
- * affect the column axis, Top/Bottom the row axis - each handle only ever touches one
- * axis, unlike the single bottom-right corner handle this replaced, which resized both
- * at once. */
-private enum class ResizeEdge { Left, Right, Top, Bottom }
-
-/**
- * Grid-snapped drag-to-resize for one edge of the selected widget - four render around a
- * selected widget (Left, Right, Top, Bottom). Right/Bottom keep the widget's own
- * gridColumn/gridRow fixed and only grow/shrink columnSpan/rowSpan, same as the original
- * corner handle. Left/Top instead keep the *opposite* edge fixed - dragging them moves
- * gridColumn/gridRow while columnSpan/rowSpan shrinks or grows to match, so the anchored
- * far edge never visibly jumps. Each handle only ever touches one axis (Left/Right:
- * columns; Top/Bottom: rows) - [onResize] always reports the full new bounds regardless,
- * with the untouched axis read fresh from [currentWidget] so it reflects whatever the
- * other three handles (or a move drag) most recently applied.
- *
- * Delta-based drag alone (grow/shrink once the finger has moved a full cell-width) cannot
- * reach the grid's true edge: right at the physical screen boundary there's no more room
- * left for the finger to keep moving, so the required delta for the last cell(s) can
- * never accumulate. [totalDrag] (tracked separately from the per-cell accumulator, never
- * reset within a gesture) gives the handle's actual on-screen position, so once that gets
- * within [EDGE_SNAP_THRESHOLD] of the grid boundary, this snaps straight to the limit
- * instead - reaching the edge only requires getting near it, not traveling the exact
- * remaining distance.
- *
- * Same stable-key + rememberUpdatedState + compute-in-onDragStart pattern as
- * PlaceholderWidgetBox - see its kdoc for why. Without this, resizing the same widget a
- * second time without deselecting in between would use whichever position/size was true
- * when the handle first appeared, not the result of the first resize.
- *
- * Deliberately a sibling of PlaceholderWidgetBox rather than nested inside it - see
- * EditWidgetsOverlay's kdoc for why.
- */
-@Composable
-private fun ResizeHandle(
-    edge: ResizeEdge,
-    widget: PlacedWidget,
-    grid: GridDimensions,
-    cellWidth: Dp,
-    cellHeight: Dp,
-    onResize: (widgetId: String, gridColumn: Int, gridRow: Int, columnSpan: Int, rowSpan: Int) -> Unit,
-    onDragStateChanged: (Boolean) -> Unit,
-    onDragEnd: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val density = LocalDensity.current
-    val hapticFeedback = LocalHapticFeedback.current
-    val currentWidget by rememberUpdatedState(widget)
-    val isColumnAxis = edge == ResizeEdge.Left || edge == ResizeEdge.Right
-    // Right/Bottom: the anchor is this widget's own near/start edge (gridColumn/gridRow),
-    // and this handle is on the far edge, growing span away from it. Left/Top: the anchor
-    // is the opposite (far) edge instead, so this handle - on the near/start edge - moves
-    // gridColumn/gridRow itself as span adjusts to compensate.
-    val isFarEdge = edge == ResizeEdge.Right || edge == ResizeEdge.Bottom
-
-    Box(
-        modifier =
-            modifier
-                .size(HANDLE_SIZE) // touch target - see HANDLE_SIZE's kdoc for why this stays full-size
-                .pointerInput(widget.id, edge, cellWidth, cellHeight, grid) {
-                    val cellWidthPx = with(density) { cellWidth.toPx() }
-                    val cellHeightPx = with(density) { cellHeight.toPx() }
-                    val cellPx = if (isColumnAxis) cellWidthPx else cellHeightPx
-                    val gridExtentCells = if (isColumnAxis) grid.columns else grid.rows
-                    val edgeSnapThresholdPx = with(density) { EDGE_SNAP_THRESHOLD.toPx() }
-
-                    var bounds = ResizeBounds(start = 0, span = 0, maxSpan = 0)
-                    var geometry = ResizeGeometry(cellPx, cellPx * gridExtentCells, edgeSnapThresholdPx, 0f, 0f)
-                    var accumulator = ResizeAccumulator(accum = 0f, totalDrag = 0f)
-
-                    detectDragGestures(
-                        onDragStart = {
-                            val liveWidget = currentWidget
-                            val start = if (isColumnAxis) liveWidget.gridColumn else liveWidget.gridRow
-                            val span = if (isColumnAxis) liveWidget.columnSpan else liveWidget.rowSpan
-                            // maxSpan means different things per anchor: for a far-edge
-                            // (Right/Bottom) handle it's the largest span can grow to; for
-                            // a near-edge (Left/Top) handle it's the anchored far edge's
-                            // fixed cell position, since span there is derived as maxSpan - start.
-                            val maxSpan = resizeMaxSpan(gridExtentCells, start, span, isFarEdge)
-                            bounds = ResizeBounds(start, span, maxSpan)
-                            geometry =
-                                geometry.copy(
-                                    initialStartPx = start * cellPx,
-                                    initialFarEdgePx = (start + span) * cellPx,
-                                )
-                            accumulator = ResizeAccumulator(accum = 0f, totalDrag = 0f)
-                            hapticFeedback.performHapticFeedback(HapticFeedbackType.LongPress)
-                            onDragStateChanged(true)
-                        },
-                        onDragEnd = {
-                            onDragStateChanged(false)
-                            onDragEnd()
-                        },
-                        onDragCancel = {
-                            onDragStateChanged(false)
-                            onDragEnd()
-                        },
-                    ) { change, dragAmount ->
-                        change.consume()
-                        val delta = if (isColumnAxis) dragAmount.x else dragAmount.y
-
-                        val step = nextResizeSpan(isFarEdge, delta, accumulator, bounds, geometry)
-                        accumulator = step.accumulator
-
-                        if (step.changed) {
-                            bounds = bounds.copy(start = step.start, span = step.span)
-                            if (step.snapped) {
-                                hapticFeedback.performHapticFeedback(HapticFeedbackType.LongPress)
-                            }
-                            val liveWidget = currentWidget
-                            onResize(
-                                widget.id,
-                                if (isColumnAxis) bounds.start else liveWidget.gridColumn,
-                                if (isColumnAxis) liveWidget.gridRow else bounds.start,
-                                if (isColumnAxis) bounds.span else liveWidget.columnSpan,
-                                if (isColumnAxis) liveWidget.rowSpan else bounds.span,
-                            )
-                        }
-                    }
-                },
-        contentAlignment = Alignment.Center,
-    ) {
-        Box(
-            modifier =
-                Modifier
-                    .size(HANDLE_DOT_SIZE)
-                    .background(MaterialTheme.colorScheme.primary, CircleShape),
-        )
-    }
-}
-
-private fun WidgetType.label(): String =
-    when (this) {
-        is WidgetType.SystemLogo -> "System Logo"
-        is WidgetType.SystemImage -> "System Image"
-        is WidgetType.SystemMedia -> mediaType.systemWidgetLabel()
-        is WidgetType.GameMedia -> mediaType.gameWidgetLabel()
-        is WidgetType.CustomImage -> "Custom Image"
-        is WidgetType.ColorBackground -> "Color Background"
-        is WidgetType.GameDescription -> "Description"
-        is WidgetType.Rating -> "Rating"
-    }
-
-/** Display label for a MediaType-backed System-canvas widget - only FanArt/Screenshots
- * ever appear there (see widgetCatalogFor), named for what they actually show: a random
- * game's art for the browsed system, not the system's own art. */
-private fun MediaType.systemWidgetLabel(): String =
-    when (this) {
-        MediaType.FanArt -> "Random Game Fanart"
-        MediaType.Screenshots -> "Random Game Screenshot"
-        else -> "System: $name"
-    }
-
-/** Display label for a MediaType-backed Game-canvas widget - plain-English names in
- * place of the raw MediaType enum constant (e.g. "Box Cover" instead of "Covers", "3D
- * Box" instead of "ThreeDBoxes"). */
-private fun MediaType.gameWidgetLabel(): String =
-    when (this) {
-        MediaType.Marquees -> "Marquee"
-        MediaType.Covers -> "Box Cover"
-        MediaType.ThreeDBoxes -> "3D Box"
-        MediaType.MixImages -> "Mix Image"
-        MediaType.Screenshots -> "Screenshot"
-        MediaType.FanArt -> "Fan Art"
-        MediaType.TitleScreens -> "Title Screen"
-        MediaType.BackCovers -> "Box Back Cover"
-        MediaType.PhysicalMedia -> "Physical Media"
-        else -> "Game: $name"
-    }
-
-/**
- * Simple list picker for widgetCatalogFor(selectedCanvas) - one dialog serves both
- * canvases, since the catalog itself is already filtered per-canvas by the caller.
- */
-@Composable
-private fun AddWidgetDialog(
-    catalog: List<WidgetType>,
-    onPick: (WidgetType) -> Unit,
-    onDismiss: () -> Unit,
-) {
-    val hapticFeedback = LocalHapticFeedback.current
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text("Add Widget") },
-        text = {
-            LazyColumn {
-                items(catalog) { widgetType ->
-                    Text(
-                        text = widgetType.label(),
-                        style = MaterialTheme.typography.bodyLarge,
-                        modifier =
-                            Modifier
-                                .fillMaxWidth()
-                                .clickable {
-                                    hapticFeedback.performHapticFeedback(HapticFeedbackType.LongPress)
-                                    onPick(widgetType)
-                                }
-                                .padding(vertical = 12.dp),
-                    )
-                }
-            }
-        },
-        confirmButton = {},
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Cancel")
-            }
-        },
-    )
-}
-
-/**
- * Content shown depends entirely on [widgetType]'s variant: image-backed types
- * (SystemLogo, SystemMedia, GameMedia) get a scale-mode choice; ColorBackground gets a
- * preset color swatch row plus an alpha slider. Neither variant needs a Confirm step -
- * [onChange] fires immediately on each selection/slider move (see
- * EditWidgetsViewModel.updateWidgetConfig's kdoc for why that's fine here), so the
- * dialog only needs a way to close, not a way to commit.
- */
-@Composable
-private fun ConfigureWidgetDialog(
-    widgetType: WidgetType,
-    onChange: (WidgetType) -> Unit,
-    onDismiss: () -> Unit,
-) {
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text("Configure Widget") },
-        text = {
-            Column(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .heightIn(max = 420.dp)
-                        .verticalScroll(rememberScrollState()),
-                verticalArrangement = Arrangement.spacedBy(12.dp),
-            ) {
-                when (widgetType) {
-                    is WidgetType.SystemLogo -> {
-                        ScaleModeConfig(current = widgetType.scaleMode) { onChange(widgetType.copy(scaleMode = it)) }
-                        LogoTransitionPicker(current = widgetType.logoTransitionMode) { onChange(widgetType.copy(logoTransitionMode = it)) }
-                        GlintConfig(enabled = widgetType.glintEnabled) { onChange(widgetType.copy(glintEnabled = it)) }
-                        ImageEffectsConfig(current = widgetType.effects) { onChange(widgetType.copy(effects = it)) }
-                    }
-
-                    is WidgetType.SystemImage -> {
-                        ScaleModeConfig(current = widgetType.scaleMode) { onChange(widgetType.copy(scaleMode = it)) }
-                        if (widgetType.supportsImageTransition && widgetType.allowsFadeTransition) {
-                            ImageTransitionPicker(
-                                current = widgetType.imageTransitionMode,
-                            ) { onChange(widgetType.copy(imageTransitionMode = it)) }
-                        }
-                        if (widgetType.supportsPanZoom) {
-                            PanZoomConfig(enabled = widgetType.panZoomEnabled) { onChange(widgetType.copy(panZoomEnabled = it)) }
-                        }
-                        ImageEffectsConfig(current = widgetType.effects) { onChange(widgetType.copy(effects = it)) }
-                    }
-
-                    is WidgetType.SystemMedia -> {
-                        ScaleModeConfig(current = widgetType.scaleMode) { onChange(widgetType.copy(scaleMode = it)) }
-                        if (widgetType.isLogoStyle) {
-                            LogoTransitionPicker(
-                                current = widgetType.logoTransitionMode,
-                            ) { onChange(widgetType.copy(logoTransitionMode = it)) }
-                            GlintConfig(enabled = widgetType.glintEnabled) { onChange(widgetType.copy(glintEnabled = it)) }
-                        } else if (widgetType.supportsImageTransition && widgetType.allowsFadeTransition) {
-                            ImageTransitionPicker(
-                                current = widgetType.imageTransitionMode,
-                            ) { onChange(widgetType.copy(imageTransitionMode = it)) }
-                        }
-                        if (widgetType.supportsPanZoom) {
-                            PanZoomConfig(enabled = widgetType.panZoomEnabled) { onChange(widgetType.copy(panZoomEnabled = it)) }
-                        }
-                        if (widgetType.supportsFallbackArtwork) {
-                            FallbackArtworkConfig(
-                                mediaType = widgetType.mediaType,
-                                current = widgetType.fallbackMediaType,
-                            ) { onChange(widgetType.copy(fallbackMediaType = it)) }
-                        }
-                        ImageEffectsConfig(current = widgetType.effects) { onChange(widgetType.copy(effects = it)) }
-                    }
-
-                    is WidgetType.GameMedia -> {
-                        ScaleModeConfig(current = widgetType.scaleMode) { onChange(widgetType.copy(scaleMode = it)) }
-                        if (widgetType.isLogoStyle) {
-                            LogoTransitionPicker(
-                                current = widgetType.logoTransitionMode,
-                            ) { onChange(widgetType.copy(logoTransitionMode = it)) }
-                            GlintConfig(enabled = widgetType.glintEnabled) { onChange(widgetType.copy(glintEnabled = it)) }
-                        } else if (widgetType.supportsImageTransition && widgetType.allowsFadeTransition) {
-                            ImageTransitionPicker(
-                                current = widgetType.imageTransitionMode,
-                            ) { onChange(widgetType.copy(imageTransitionMode = it)) }
-                        }
-                        if (widgetType.supportsPanZoom) {
-                            PanZoomConfig(enabled = widgetType.panZoomEnabled) { onChange(widgetType.copy(panZoomEnabled = it)) }
-                        }
-                        if (widgetType.supportsFallbackArtwork) {
-                            FallbackArtworkConfig(
-                                mediaType = widgetType.mediaType,
-                                current = widgetType.fallbackMediaType,
-                            ) { onChange(widgetType.copy(fallbackMediaType = it)) }
-                        }
-                        ImageEffectsConfig(current = widgetType.effects) { onChange(widgetType.copy(effects = it)) }
-                    }
-
-                    is WidgetType.CustomImage -> {
-                        CustomImageConfig(current = widgetType, onChange = onChange)
-                        ScaleModeConfig(current = widgetType.scaleMode) { onChange(widgetType.copy(scaleMode = it)) }
-                        if (widgetType.supportsImageTransition && widgetType.allowsFadeTransition) {
-                            ImageTransitionPicker(
-                                current = widgetType.imageTransitionMode,
-                            ) { onChange(widgetType.copy(imageTransitionMode = it)) }
-                        }
-                        if (widgetType.supportsPanZoom) {
-                            PanZoomConfig(enabled = widgetType.panZoomEnabled) { onChange(widgetType.copy(panZoomEnabled = it)) }
-                        }
-                        ImageEffectsConfig(current = widgetType.effects) { onChange(widgetType.copy(effects = it)) }
-                    }
-
-                    is WidgetType.ColorBackground ->
-                        ColorBackgroundConfig(current = widgetType, onChange = onChange)
-
-                    is WidgetType.GameDescription ->
-                        GameDescriptionConfig(current = widgetType, onChange = onChange)
-
-                    is WidgetType.Rating ->
-                        RatingConfig(current = widgetType, onChange = onChange)
-                }
-            }
-        },
-        confirmButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Done")
-            }
-        },
-    )
-}
-
-/** Shows the currently-picked file's name (or a prompt if none picked yet, e.g. this
- * widget was just added and the picker was cancelled) and a button to (re)launch the
- * same document picker used on add - see addCustomImageLauncher in EditWidgetsOverlay. */
-@Composable
-private fun CustomImageConfig(
-    current: WidgetType.CustomImage,
-    onChange: (WidgetType.CustomImage) -> Unit,
-) {
-    val launcher =
-        rememberLauncherForActivityResult(ActivityResultContracts.OpenDocument()) { uri ->
-            uri?.let(SafPathResolver::resolveDocumentPath)?.let { path -> onChange(current.copy(path = path)) }
-        }
-
-    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-        Text(text = "Image", style = MaterialTheme.typography.titleSmall)
-        Text(
-            text = if (current.path.isNotBlank()) current.path.substringAfterLast('/') else "No image chosen",
-            style = MaterialTheme.typography.bodyMedium,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-        )
-        TextButton(onClick = { launcher.launch(arrayOf("image/*")) }) {
-            Text(if (current.path.isNotBlank()) "Change Image" else "Choose Image")
-        }
-    }
-}
-
-private fun ScaleMode.displayLabel(): String =
-    when (this) {
-        ScaleMode.Fit -> "Contain"
-        ScaleMode.Fill -> "Cover"
-    }
-
-private fun ScaleMode.icon(): ImageVector =
-    when (this) {
-        ScaleMode.Fit -> Icons.Filled.CropFree // whole image visible, letterboxed
-        ScaleMode.Fill -> Icons.Filled.Crop // cropped to fill the frame
-    }
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun ScaleModeConfig(
-    current: ScaleMode,
-    onSelect: (ScaleMode) -> Unit,
-) {
-    Column {
-        Text(text = "Image Scaling", style = MaterialTheme.typography.titleSmall)
-        Spacer(modifier = Modifier.height(8.dp))
-        SingleChoiceSegmentedButtonRow(modifier = Modifier.fillMaxWidth()) {
-            ScaleMode.entries.forEachIndexed { index, mode ->
-                SegmentedButton(
-                    selected = mode == current,
-                    onClick = { onSelect(mode) },
-                    shape = SegmentedButtonDefaults.itemShape(index = index, count = ScaleMode.entries.size),
-                    icon = {
-                        SegmentedButtonDefaults.Icon(active = mode == current) {
-                            Icon(
-                                imageVector = mode.icon(),
-                                contentDescription = null,
-                                modifier = Modifier.size(SegmentedButtonDefaults.IconSize),
-                            )
-                        }
-                    },
-                    label = { Text(mode.displayLabel()) },
-                )
-            }
-        }
-    }
-}
-
-/**
- * Blur + darken controls for image-backed widget types (SystemLogo, SystemMedia,
- * GameMedia). Both scale as simple 0f..1f sliders - see ImageEffects' kdoc for why, and
- * applyBlurEffect/DarkenOverlay in WidgetCanvas.kt for how these map to rendering.
- * Darken is a flat black overlay rather than a tint color - its main use is muting a
- * busy background image so a logo/widget placed on top reads more clearly.
- */
-@Composable
-private fun ImageEffectsConfig(
-    current: ImageEffects,
-    onChange: (ImageEffects) -> Unit,
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-        Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-            Text(text = "Blur: ${(current.blurAmount * 100).roundToInt()}%", style = MaterialTheme.typography.titleSmall)
-            Slider(
-                value = current.blurAmount,
-                onValueChange = { onChange(current.copy(blurAmount = it)) },
-                valueRange = 0f..1f,
-            )
-        }
-        Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-            Text(text = "Darken: ${(current.darkenAmount * 100).roundToInt()}%", style = MaterialTheme.typography.titleSmall)
-            Slider(
-                value = current.darkenAmount,
-                onValueChange = { onChange(current.copy(darkenAmount = it)) },
-                valueRange = 0f..1f,
-            )
-        }
-    }
-}
-
-/**
- * Only shown when [WidgetType.supportsPanZoom] is true for the widget being configured -
- * eligible image-backed types with ScaleMode.Fill (see its kdoc for why Fit is excluded).
- * See PanZoomImage.kt for the animation this toggle drives.
- */
-@Composable
-private fun PanZoomConfig(
-    enabled: Boolean,
-    onChange: (Boolean) -> Unit,
-) {
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Column(modifier = Modifier.weight(1f)) {
-            Text(text = "Pan & Zoom", style = MaterialTheme.typography.titleSmall)
-            Text(
-                text = "Slowly zooms and pans across the image while it's displayed",
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-        }
-        Switch(checked = enabled, onCheckedChange = onChange)
-    }
-}
-
-/**
- * Only shown when [WidgetType.supportsFallbackArtwork] is true - lets a Fan Art,
- * Screenshot, or 3D Box widget substitute a different MediaType (or nothing) when its
- * own primary artwork isn't available for the currently browsed system/game - see
- * [MediaType.fallbackMediaTypeOptions] for which choices apply to which [mediaType], and
- * resolveMediaWidgetContent for where this is actually applied at render time.
- */
-@Composable
-private fun FallbackArtworkConfig(
-    mediaType: MediaType,
-    current: MediaType?,
-    onSelected: (MediaType?) -> Unit,
-) {
-    val options = mediaType.fallbackMediaTypeOptions()
-    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-        Text(text = "Fallback Artwork", style = MaterialTheme.typography.titleSmall)
-        SingleChoiceSegmentedButtonRow(modifier = Modifier.fillMaxWidth()) {
-            options.forEachIndexed { index, option ->
-                SegmentedButton(
-                    selected = option == current,
-                    onClick = { onSelected(option) },
-                    shape = SegmentedButtonDefaults.itemShape(index = index, count = options.size),
-                    label = { Text(option.fallbackArtworkDisplayLabel()) },
-                )
-            }
-        }
-    }
-}
-
-/** Plain-English label for a Fallback Artwork choice - deliberately distinct from
- * MediaType.gameWidgetLabel()/systemWidgetLabel() (those describe what a whole widget
- * shows, e.g. "Random Game Fanart"; this describes a fallback target on its own, e.g.
- * just "Fan Art"). `null` is the explicit "None" choice. */
-private fun MediaType?.fallbackArtworkDisplayLabel(): String =
-    when (this) {
-        null -> "None"
-        MediaType.FanArt -> "Fan Art"
-        MediaType.Screenshots -> "Screenshot"
-        MediaType.Covers -> "Box Cover"
-        else -> name
-    }
-
-/**
- * Only shown when [WidgetType.supportsImageTransition] AND [WidgetType.allowsFadeTransition]
- * are both true for the widget being configured (opaque/backdrop-style image widgets, not
- * logo-style or permanently-instant box-art-style GameMedia - see supportsImageTransition's
- * kdoc). allowsFadeTransition is false whenever the widget has ScaleMode.Fit selected, in
- * which case Fade is the only entry that would matter and None is fixed anyway - hidden
- * entirely rather than shown as a locked one-option row, same reasoning as
- * [WidgetType.supportsPanZoom] hiding rather than graying out.
- */
-@Composable
-private fun ImageTransitionPicker(
-    current: ImageTransitionMode,
-    onSelected: (ImageTransitionMode) -> Unit,
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-        Text(text = "Image Transitions", style = MaterialTheme.typography.titleSmall)
-        SingleChoiceSegmentedButtonRow(modifier = Modifier.fillMaxWidth()) {
-            ImageTransitionMode.entries.forEachIndexed { index, mode ->
-                SegmentedButton(
-                    selected = mode == current,
-                    onClick = { onSelected(mode) },
-                    shape = SegmentedButtonDefaults.itemShape(index = index, count = ImageTransitionMode.entries.size),
-                    icon = {
-                        SegmentedButtonDefaults.Icon(active = mode == current) {
-                            Icon(
-                                imageVector = mode.icon(),
-                                contentDescription = null,
-                                modifier = Modifier.size(SegmentedButtonDefaults.IconSize),
-                            )
-                        }
-                    },
-                    label = { Text(mode.displayLabel()) },
-                )
-            }
-        }
-    }
-}
-
-private fun ImageTransitionMode.icon(): ImageVector =
-    when (this) {
-        ImageTransitionMode.None -> Icons.Filled.Block
-        ImageTransitionMode.Fade -> Icons.Filled.BlurOn
-    }
-
-private fun ImageTransitionMode.displayLabel(): String =
-    when (this) {
-        ImageTransitionMode.None -> "None"
-        ImageTransitionMode.Fade -> "Fade"
-    }
-
-/**
- * Shown for every logo-style widget (SystemLogo, or SystemMedia/GameMedia with mediaType
- * Marquees - see [WidgetType.isLogoStyle]). Deliberately no Fade option - see
- * AnimatedLogoImage's kdoc for why (double-exposure on overlapping transparent images).
- */
-@Composable
-private fun LogoTransitionPicker(
-    current: LogoTransitionMode,
-    onSelected: (LogoTransitionMode) -> Unit,
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-        Text(text = "Logo Transitions", style = MaterialTheme.typography.titleSmall)
-        SingleChoiceSegmentedButtonRow(modifier = Modifier.fillMaxWidth()) {
-            LogoTransitionMode.entries.forEachIndexed { index, mode ->
-                SegmentedButton(
-                    selected = mode == current,
-                    onClick = { onSelected(mode) },
-                    shape = SegmentedButtonDefaults.itemShape(index = index, count = LogoTransitionMode.entries.size),
-                    icon = {
-                        SegmentedButtonDefaults.Icon(active = mode == current) {
-                            Icon(
-                                imageVector = mode.icon(),
-                                contentDescription = null,
-                                modifier = Modifier.size(SegmentedButtonDefaults.IconSize),
-                            )
-                        }
-                    },
-                    label = { Text(mode.displayLabel()) },
-                )
-            }
-        }
-    }
-}
-
-private fun LogoTransitionMode.icon(): ImageVector =
-    when (this) {
-        LogoTransitionMode.None -> Icons.Filled.Block
-        LogoTransitionMode.Slide -> Icons.Filled.ArrowForward
-        LogoTransitionMode.Scale -> Icons.Filled.ZoomIn
-    }
-
-private fun LogoTransitionMode.displayLabel(): String =
-    when (this) {
-        LogoTransitionMode.None -> "None"
-        LogoTransitionMode.Slide -> "Slide"
-        LogoTransitionMode.Scale -> "Scale"
-    }
-
-/**
- * Shown alongside [LogoTransitionPicker] for the same logo-style widget types -
- * independent of [LogoTransitionMode], a periodic ambient shimmer orthogonal to whichever
- * entrance-transition mode is selected above (see AnimatedLogoImage's glint loop).
- */
-@Composable
-private fun GlintConfig(
-    enabled: Boolean,
-    onChange: (Boolean) -> Unit,
-) {
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Column(modifier = Modifier.weight(1f)) {
-            Text(text = "Logo Glint", style = MaterialTheme.typography.titleSmall)
-            Text(
-                text = "A periodic light sweep across this widget",
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-        }
-        Switch(checked = enabled, onCheckedChange = onChange)
-    }
-}
-
-/** Fixed preset palette - a full custom color picker (hue/saturation wheel) is more
- * than this widget type needs for a first pass; these presets cover the common cases
- * (tint/darken behind other widgets) with a much simpler UI. */
-private val COLOR_PRESETS =
-    listOf(
-        // black
-        0xFF000000L,
-        // white
-        0xFFFFFFFFL,
-        // gray
-        0xFF9E9E9EL,
-        // red
-        0xFFF44336L,
-        // blue
-        0xFF2196F3L,
-        // green
-        0xFF4CAF50L,
-        // yellow
-        0xFFFFEB3BL,
-        // purple
-        0xFF9C27B0L,
-    )
-
-@Composable
-private fun ColorBackgroundConfig(
-    current: WidgetType.ColorBackground,
-    onChange: (WidgetType.ColorBackground) -> Unit,
-) {
-    val hapticFeedback = LocalHapticFeedback.current
-    Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-        Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-            Text(text = "Color", style = MaterialTheme.typography.titleSmall)
-            Row(modifier = Modifier.horizontalScroll(rememberScrollState())) {
-                COLOR_PRESETS.forEach { colorArgb ->
-                    Box(
-                        modifier =
-                            Modifier
-                                .padding(4.dp)
-                                .size(36.dp)
-                                .background(Color(colorArgb), CircleShape)
-                                .then(
-                                    if (colorArgb == current.colorArgb) {
-                                        Modifier.border(2.dp, MaterialTheme.colorScheme.primary, CircleShape)
-                                    } else {
-                                        Modifier
-                                    },
-                                )
-                                .clickable {
-                                    hapticFeedback.performHapticFeedback(HapticFeedbackType.LongPress)
-                                    onChange(current.copy(colorArgb = colorArgb))
-                                },
-                    )
-                }
-            }
-            HexColorInput(current = current.colorArgb) { onChange(current.copy(colorArgb = it)) }
-        }
-        Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-            Text(text = "Transparency: ${(current.alpha * 100).roundToInt()}%", style = MaterialTheme.typography.titleSmall)
-            Slider(
-                value = current.alpha,
-                onValueChange = { onChange(current.copy(alpha = it)) },
-                valueRange = 0f..1f,
-            )
-        }
-    }
-}
-
-@Composable
-private fun GameDescriptionConfig(
-    current: WidgetType.GameDescription,
-    onChange: (WidgetType.GameDescription) -> Unit,
-) {
-    val hapticFeedback = LocalHapticFeedback.current
-    Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-        Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-            Text(text = "Text Size: ${current.fontSizeSp.roundToInt()}sp", style = MaterialTheme.typography.titleSmall)
-            Slider(
-                value = current.fontSizeSp,
-                onValueChange = { onChange(current.copy(fontSizeSp = it)) },
-                valueRange = 10f..36f,
-            )
-        }
-        Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-            Text(text = "Text Color", style = MaterialTheme.typography.titleSmall)
-            Row(modifier = Modifier.horizontalScroll(rememberScrollState())) {
-                COLOR_PRESETS.forEach { colorArgb ->
-                    Box(
-                        modifier =
-                            Modifier
-                                .padding(4.dp)
-                                .size(36.dp)
-                                .background(Color(colorArgb), CircleShape)
-                                .then(
-                                    if (colorArgb == current.textColorArgb) {
-                                        Modifier.border(2.dp, MaterialTheme.colorScheme.primary, CircleShape)
-                                    } else {
-                                        Modifier
-                                    },
-                                )
-                                .clickable {
-                                    hapticFeedback.performHapticFeedback(HapticFeedbackType.LongPress)
-                                    onChange(current.copy(textColorArgb = colorArgb))
-                                },
-                    )
-                }
-            }
-            HexColorInput(current = current.textColorArgb) { onChange(current.copy(textColorArgb = it)) }
-        }
-        Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-            Text(text = "Background Color", style = MaterialTheme.typography.titleSmall)
-            Row(modifier = Modifier.horizontalScroll(rememberScrollState())) {
-                COLOR_PRESETS.forEach { colorArgb ->
-                    Box(
-                        modifier =
-                            Modifier
-                                .padding(4.dp)
-                                .size(36.dp)
-                                .background(Color(colorArgb), CircleShape)
-                                .then(
-                                    if (colorArgb == current.backgroundColorArgb) {
-                                        Modifier.border(2.dp, MaterialTheme.colorScheme.primary, CircleShape)
-                                    } else {
-                                        Modifier
-                                    },
-                                )
-                                .clickable {
-                                    hapticFeedback.performHapticFeedback(HapticFeedbackType.LongPress)
-                                    onChange(current.copy(backgroundColorArgb = colorArgb))
-                                },
-                    )
-                }
-            }
-            HexColorInput(current = current.backgroundColorArgb) { onChange(current.copy(backgroundColorArgb = it)) }
-        }
-        Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-            Text(
-                text = "Background Transparency: ${(current.backgroundAlpha * 100).roundToInt()}%",
-                style = MaterialTheme.typography.titleSmall,
-            )
-            Slider(
-                value = current.backgroundAlpha,
-                onValueChange = { onChange(current.copy(backgroundAlpha = it)) },
-                valueRange = 0f..1f,
-            )
-        }
-    }
-}
-
-/**
- * What a Rating widget shows for a game with no <rating> at all in gamelist.xml - see
- * [WidgetType.Rating.noRatingBehavior]'s kdoc for why this is distinct from a rating of
- * exactly 0.
- */
-private fun NoRatingBehavior.displayLabel(): String =
-    when (this) {
-        NoRatingBehavior.Hide -> "Hide Widget"
-        NoRatingBehavior.ShowEmptyStars -> "Show Empty"
-    }
-
-@Composable
-private fun RatingConfig(
-    current: WidgetType.Rating,
-    onChange: (WidgetType.Rating) -> Unit,
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-        Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-            Text(text = "No Rating Behavior", style = MaterialTheme.typography.titleSmall)
-            SingleChoiceSegmentedButtonRow(modifier = Modifier.fillMaxWidth()) {
-                NoRatingBehavior.entries.forEachIndexed { index, behavior ->
-                    SegmentedButton(
-                        selected = behavior == current.noRatingBehavior,
-                        onClick = { onChange(current.copy(noRatingBehavior = behavior)) },
-                        shape = SegmentedButtonDefaults.itemShape(index = index, count = NoRatingBehavior.entries.size),
-                        label = { Text(behavior.displayLabel()) },
-                    )
-                }
-            }
-        }
-        ColorSwatchPicker(
-            label = "Star Fill Color",
-            current = current.filledColorArgb,
-        ) { onChange(current.copy(filledColorArgb = it)) }
-        ColorSwatchPicker(
-            label = "Star Outline Color",
-            current = current.outlineColorArgb,
-        ) { onChange(current.copy(outlineColorArgb = it)) }
-        ColorSwatchPicker(
-            label = "Background Color",
-            current = current.backgroundColorArgb,
-        ) { onChange(current.copy(backgroundColorArgb = it)) }
-        Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-            Text(
-                text = "Background Transparency: ${(current.backgroundAlpha * 100).roundToInt()}%",
-                style = MaterialTheme.typography.titleSmall,
-            )
-            Slider(
-                value = current.backgroundAlpha,
-                onValueChange = { onChange(current.copy(backgroundAlpha = it)) },
-                valueRange = 0f..1f,
-            )
-        }
-    }
-}
-
-/**
- * A labeled preset-swatch row + [HexColorInput], the exact block ColorBackgroundConfig/
- * GameDescriptionConfig each inline once or twice - extracted here since RatingConfig
- * needs it three times (fill, outline, background) in one dialog.
- */
-@Composable
-private fun ColorSwatchPicker(
-    label: String,
-    current: Long,
-    onChange: (Long) -> Unit,
-) {
-    val hapticFeedback = LocalHapticFeedback.current
-    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-        Text(text = label, style = MaterialTheme.typography.titleSmall)
-        Row(modifier = Modifier.horizontalScroll(rememberScrollState())) {
-            COLOR_PRESETS.forEach { colorArgb ->
-                Box(
-                    modifier =
-                        Modifier
-                            .padding(4.dp)
-                            .size(36.dp)
-                            .background(Color(colorArgb), CircleShape)
-                            .then(
-                                if (colorArgb == current) {
-                                    Modifier.border(2.dp, MaterialTheme.colorScheme.primary, CircleShape)
-                                } else {
-                                    Modifier
-                                },
-                            )
-                            .clickable {
-                                hapticFeedback.performHapticFeedback(HapticFeedbackType.LongPress)
-                                onChange(colorArgb)
-                            },
-                )
-            }
-        }
-        HexColorInput(current = current, onValidHex = onChange)
-    }
-}
-
-/**
- * Free-form hex entry alongside the preset swatches. [current]'s own alpha channel is
- * always opaque (0xFF...) - the presets are stored that way and transparency is handled
- * separately by ColorBackground.alpha - so this only ever reads/writes the RGB portion.
- *
- * Local [text] is keyed on [current] rather than plain `remember { }`: this means
- * picking a preset (or another external change) reformats the field to match, while
- * typing an in-progress/invalid hex string is left alone across recompositions since
- * [current] doesn't change until a full valid 6-digit value is entered. onValidHex only
- * fires once the text parses cleanly, so partial input never produces a garbage color.
- */
-@Composable
-private fun HexColorInput(
-    current: Long,
-    onValidHex: (Long) -> Unit,
-) {
-    var text by remember(current) { mutableStateOf(current.toHexRgbString()) }
-    val isValid = parseHexColor(text) != null
-
-    OutlinedTextField(
-        value = text,
-        onValueChange = { newText ->
-            text = newText
-            parseHexColor(newText)?.let(onValidHex)
-        },
-        label = { Text("Hex code") },
-        leadingIcon = { Text(text = "#", style = MaterialTheme.typography.bodyLarge) },
-        isError = !isValid,
-        supportingText = { if (!isValid) Text("Enter a 6-digit hex code, e.g. 3A7BD5") },
-        singleLine = true,
-        modifier = Modifier.fillMaxWidth(),
-    )
-}
-
-internal fun Long.toHexRgbString(): String = String.format(Locale.ROOT, "%06X", this and 0xFFFFFF)
-
-internal fun parseHexColor(text: String): Long? {
-    val cleaned = text.removePrefix("#").trim()
-    if (cleaned.length != 6 || cleaned.any { it !in HEX_CHARS }) return null
-    return cleaned.toLongOrNull(16)?.let { 0xFF000000L or it }
-}
-
-private val HEX_CHARS = "0123456789abcdefABCDEF".toSet()
